replace BotBowlEnv with NewBotBowlEnv, (destroys env.py history)
