+++ botbowl/ai/env_render.py
@@ -1,4 +1,4 @@
-#from .new_env import NewBotBowlEnv
+#from .new_env import BotBowlEnv
 import tkinter as tk
 import math
 
@@ -21,7 +21,7 @@ class EnvRenderer:
     wing = '#55aa55'
     scrimmage = '#338833'
 
-    def __init__(self, env: 'NewBotBowlEnv', feature_layers=False):
+    def __init__(self, env: 'BotBowlEnv', feature_layers=False):
         self.env = env
         self.layers = env.env_conf.layers
         self.root = tk.Tk()
